// FE/src/api/models/auditLog.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * SIGREF API
 * API de Auditoría para Sistema de Gestión de Referencias FHIR
 * OpenAPI spec version: 1.0.0
 */
import type { AuditLogFiltersUsed } from "./auditLogFiltersUsed";
import type { AuditLogAdditionalInfo } from "./auditLogAdditionalInfo";

/**
 * Modelo de auditoría de capa API (Sustituye al antiguo modelo basado en FHIR AuditEvent)
 */
export interface AuditLog {
  /** @nullable */
  id?: string | null;
  /** @nullable */
  traceId?: string | null;
  /** @nullable */
  action?: string | null;
  /** @nullable */
  resourceType?: string | null;
  /** @nullable */
  resourceId?: string | null;
  /** @nullable */
  userId?: string | null;
  /** @nullable */
  userName?: string | null;
  /** @nullable */
  ipAddress?: string | null;
  timestamp?: string;
  /** @nullable */
  endpoint?: string | null;
  /** @nullable */
  httpMethod?: string | null;
  statusCode?: number;
  /**
   * Lista de las rutas JSON que fueron modificadas (POST/PUT/PATCH).
   * @nullable
   */
  affectedKeys?: string[] | null;
  /**
   * Filtros o query parameters utilizados en consultas (GET).
   * @nullable
   */
  filtersUsed?: AuditLogFiltersUsed;
  success?: boolean;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  additionalInfo?: AuditLogAdditionalInfo;
}